car_numbers/requirements.txt
Django==4.1.5
gunicorn==20.1.0
djoser==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.4
psycopg2-binary==2.9.5